+++ src/egon/data/datasets/electrical_neighbours.py
@@ -1,25 +1,27 @@
 """The central module containing all code dealing with electrical neighbours
 """
 
+from os import path
+from pathlib import Path
+import datetime
+import logging
+import os.path
 import zipfile
 
-import entsoe
-import requests
-import logging
-
-import geopandas as gpd
-import pandas as pd
 from shapely.geometry import LineString
 from sqlalchemy.orm import sessionmaker
+import entsoe
+import geopandas as gpd
+import pandas as pd
+import requests
 
+from egon.data import config, db, logger
+from egon.data.datasets import Dataset, wrapped_partial
+from egon.data.datasets.fill_etrago_gen import add_marginal_costs
+from egon.data.datasets.fix_ehv_subnetworks import select_bus_id
+from egon.data.datasets.scenario_parameters import get_sector_parameters
 import egon.data.datasets.etrago_setup as etrago
 import egon.data.datasets.scenario_parameters.parameters as scenario_parameters
-from egon.data import config, db
-from egon.data.datasets import Dataset
-from egon.data.datasets.fix_ehv_subnetworks import select_bus_id
-from egon.data.datasets.fill_etrago_gen import add_marginal_costs
-from egon.data.datasets.scenario_parameters import get_sector_parameters
-from os import path
 
 
 def get_cross_border_buses(scenario, sources):
@@ -248,7 +250,11 @@ def buses(scenario, sources, targets):
     central_buses.scn_name = scenario
 
     # Insert all central buses for eGon2035
-    if scenario in ["eGon2035", "status2019"]:
+    if scenario in [
+        "eGon2035",
+        "status2019",
+        "status2023",
+    ]:  # TODO: status2023 this is hardcoded shit
         central_buses.to_postgis(
             targets["buses"]["table"],
             schema=targets["buses"]["schema"],
@@ -588,7 +594,7 @@ def foreign_dc_lines(scenario, sources, targets, central_buses):
               AND country != 'DE')
         """
     )
-    capital_cost = get_sector_parameters("electricity", "eGon2035")[
+    capital_cost = get_sector_parameters("electricity", scenario)[
         "capital_cost"
     ]
 
@@ -1108,9 +1114,9 @@ def insert_storage(capacities):
 
     for x in parameters:
         store.loc[store["carrier"] == "battery", x] = parameters_battery[x]
-        store.loc[
-            store["carrier"] == "pumped_hydro", x
-        ] = parameters_pumped_hydro[x]
+        store.loc[store["carrier"] == "pumped_hydro", x] = (
+            parameters_pumped_hydro[x]
+        )
 
     # insert data
     session = sessionmaker(bind=db.engine())()
@@ -1231,9 +1237,9 @@ def tyndp_demand():
     ]
     # Assign etrago bus_id to TYNDP nodes
     buses = pd.DataFrame({"nodes": nodes})
-    buses.loc[
-        buses[buses.nodes.isin(map_buses.keys())].index, "nodes"
-    ] = buses[buses.nodes.isin(map_buses.keys())].nodes.map(map_buses)
+    buses.loc[buses[buses.nodes.isin(map_buses.keys())].index, "nodes"] = (
+        buses[buses.nodes.isin(map_buses.keys())].nodes.map(map_buses)
+    )
     buses.loc[:, "bus"] = (
         get_foreign_bus_id(scenario="eGon2035")
         .loc[buses.loc[:, "nodes"]]
@@ -1300,12 +1306,26 @@ def tyndp_demand():
         session.commit()
 
 
+def get_entsoe_token():
+    """Check for token in home dir. If not exists, check in working dir"""
+    token_path = path.join(path.expanduser("~"), ".entsoe-token")
+    if not os.path.isfile(token_path):
+        logger.info(
+            f"Token file not found at {token_path}. Will check in working directory."
+        )
+        token_path = Path(".entsoe-token")
+        if os.path.isfile(token_path):
+            logger.info(f"Token found at {token_path}")
+    entsoe_token = open(token_path, "r").read(36)
+    if entsoe_token is None:
+        raise FileNotFoundError("No entsoe-token found.")
+    return entsoe_token
+
+
 def entsoe_historic_generation_capacities(
     year_start="20190101", year_end="20200101"
 ):
-    entsoe_token = open(
-        path.join(path.expanduser("~"), ".entsoe-token"), "r"
-    ).read(36)
+    entsoe_token = get_entsoe_token()
     client = entsoe.EntsoePandasClient(api_key=entsoe_token)
 
     start = pd.Timestamp(year_start, tz="Europe/Brussels")
@@ -1327,7 +1347,12 @@ def entsoe_historic_generation_capacities(
         "SE",
         "GB",
     ]
-
+    # No GB data after Brexit
+    if int(year_start[:4]) > 2021:
+        logger.warning(
+            "No GB data after Brexit. GB is dropped from entsoe query!"
+        )
+        countries = [c for c in countries if c != "GB"]
     # todo: define wanted countries
 
     not_retrieved = []
@@ -1338,29 +1363,53 @@ def entsoe_historic_generation_capacities(
         else:
             kwargs = dict(start=start, end=end)
         try:
-            dfs.append(
-                client.query_installed_generation_capacity(country, **kwargs)
+            country_data = client.query_installed_generation_capacity(
+                country, **kwargs
             )
-
+            dfs.append(country_data)
         except (entsoe.exceptions.NoMatchingDataError, requests.HTTPError):
+            logger.warning(
+                f"Data for country: {country} could not be retrieved."
+            )
             not_retrieved.append(country)
             pass
 
-    if not_retrieved:
-        logging.warning(
-            f"Data for country (-ies) {', '.join(not_retrieved)} could not be retrieved."
-        )
-    df = pd.concat(dfs)
-    df["country"] = countries
-    df.set_index("country", inplace=True)
-    df.fillna(0, inplace=True)
-    return df
+    if dfs:
+        df = pd.concat(dfs)
+        df["country"] = [c for c in countries if c not in not_retrieved]
+        df.set_index("country", inplace=True)
+        if int(year_start[:4]) == 2023:
+            # https://www.bmreports.com/bmrs/?q=foregeneration/capacityaggregated
+            # could probably somehow be automised
+            # https://www.elexonportal.co.uk/category/view/178
+            # in MW
+            installed_capacity_gb = pd.Series(
+                {
+                    "Biomass": 4438,
+                    "Fossil Gas": 37047,
+                    "Fossil Hard coal": 1491,
+                    "Hydro Pumped Storage": 5603,
+                    "Hydro Run-of-river and poundage": 2063,
+                    "Nuclear": 4950,
+                    "Other": 3313,
+                    "Other renewable": 1462,
+                    "Solar": 14518,
+                    "Wind Offshore": 13038,
+                    "Wind Onshore": 13907,
+                },
+                name="GB",
+            )
+            df = pd.concat([df.T, installed_capacity_gb], axis=1).T
+            logger.info("Manually added generation capacities for GB 2023.")
+            not_retrieved = [c for c in not_retrieved if c != "GB"]
+        df.fillna(0, inplace=True)
+    else:
+        df = pd.DataFrame()
+    return df, not_retrieved
 
 
 def entsoe_historic_demand(year_start="20190101", year_end="20200101"):
-    entsoe_token = open(
-        path.join(path.expanduser("~"), ".entsoe-token"), "r"
-    ).read(36)
+    entsoe_token = get_entsoe_token()
     client = entsoe.EntsoePandasClient(api_key=entsoe_token)
 
     start = pd.Timestamp(year_start, tz="Europe/Brussels")
@@ -1407,17 +1456,18 @@ def entsoe_historic_demand(year_start="20190101", year_end="20200101"):
             dfs.append(country_data)
         except (entsoe.exceptions.NoMatchingDataError, requests.HTTPError):
             not_retrieved.append(country)
+            logger.warning(
+                f"Data for country: {country} could not be retrieved."
+            )
             pass
-    if not_retrieved:
-        logging.warning(
-            f"Data for country (-ies) {', '.join(not_retrieved)} could not be retrieved."
-        )
 
-    df = pd.concat(dfs, axis=1)
-    df.columns = countries
-    df.index = pd.date_range(year_start, periods=8760, freq="H")
-
-    return df
+    if dfs:
+        df = pd.concat(dfs, axis=1)
+        df.columns = [c for c in countries if c not in not_retrieved]
+        df.index = pd.date_range(year_start, periods=8760, freq="H")
+    else:
+        df = pd.DataFrame()
+    return df, not_retrieved
 
 
 def map_carriers_entsoe():
@@ -1451,7 +1501,7 @@ def map_carriers_entsoe():
     }
 
 
-def entsoe_to_bus_etrago():
+def entsoe_to_bus_etrago(scn_name):
     map_entsoe = pd.Series(
         {
             "LU": "LU00",
@@ -1470,9 +1520,19 @@ def entsoe_to_bus_etrago():
         }
     )
 
-    for_bus = get_foreign_bus_id(scenario="status2019")
+    for_bus = get_foreign_bus_id(scenario=scn_name)
 
     return map_entsoe.map(for_bus)
+
+
+def save_entsoe_data(df: pd.DataFrame, file_path: Path):
+    os.makedirs(file_path.parent, exist_ok=True)
+    if not df.empty:
+        df.to_csv(file_path, index_label="Index")
+        logger.info(
+            f"Saved entsoe data for {file_path.stem} "
+            f"to {file_path.parent} for countries: {df.index}"
+        )
 
 
 def insert_generators_sq(scn_name="status2019"):
@@ -1492,16 +1552,44 @@ def insert_generators_sq(scn_name="status2019"):
     None.
 
     """
-    try:
-        gen_sq = entsoe_historic_generation_capacities()
-    except:
-        logging.warning(
-            """Generation data from entsoe could not be retrieved.
-                        Backup data is used instead"""
-        )
-        gen_sq = pd.read_csv(
+
+    if scn_name == "status2019":
+        year_start_end = {"year_start": "20190101", "year_end": "20200101"}
+    elif scn_name == "status2023":
+        year_start_end = {"year_start": "20230101", "year_end": "20240101"}
+    else:
+        raise ValueError("No valid scenario name!")
+
+    gen_sq, not_retrieved = entsoe_historic_generation_capacities(
+        **year_start_end
+    )
+
+    if not_retrieved:
+        logger.warning("Generation data from entsoe could not be retrieved.")
+        # check for generation backup from former runs
+        file_path = Path(
+            "./", "entsoe_data", f"gen_entsoe_{scn_name}.csv"
+        ).resolve()
+        if os.path.isfile(file_path):
+            gen_sq_backup = pd.read_csv(file_path, index_col="Index")
+            # check for missing columns in backup (former runs)
+            c_backup = [c for c in gen_sq_backup.columns if c in not_retrieved]
+            # remove columns, if found in backup
+            not_retrieved = [c for c in not_retrieved if c not in c_backup]
+            if c_backup:
+                gen_sq = pd.concat(
+                    [gen_sq, gen_sq_backup.loc[:, c_backup]], axis=1
+                )
+                logger.info(f"Appended data from former runs for {c_backup}")
+        save_entsoe_data(gen_sq, file_path=file_path)
+
+    if not_retrieved:
+        logger.warning("Backup data from 2019 is used instead.")
+        gen_sq_backup = pd.read_csv(
             "data_bundle_egon_data/entsoe/gen_entsoe.csv", index_col="Index"
         )
+
+        gen_sq = pd.concat([gen_sq, gen_sq_backup.loc[not_retrieved]], axis=1)
 
     targets = config.datasets()["electrical_neighbours"]["targets"]
     # Delete existing data
@@ -1531,7 +1619,7 @@ def insert_generators_sq(scn_name="status2019"):
         AND scn_name = '{scn_name}'
         """
     )
-    entsoe_to_bus = entsoe_to_bus_etrago()
+    entsoe_to_bus = entsoe_to_bus_etrago(scn_name)
     carrier_entsoe = map_carriers_entsoe()
     gen_sq = gen_sq.groupby(axis=1, by=carrier_entsoe).sum()
 
@@ -1650,15 +1738,47 @@ def insert_loads_sq(scn_name="status2019"):
     """
     sources = config.datasets()["electrical_neighbours"]["sources"]
     targets = config.datasets()["electrical_neighbours"]["targets"]
-    try:
-        load_sq = entsoe_historic_demand()
-    except:
-        logging.warning(
-            """Demand data from entsoe could not be retrieved.
-                        Backup data is used instead"""
+
+    if scn_name == "status2019":
+        year_start_end = {"year_start": "20190101", "year_end": "20200101"}
+    elif scn_name == "status2023":
+        year_start_end = {"year_start": "20230101", "year_end": "20240101"}
+    else:
+        raise ValueError("No valid scenario name!")
+
+    load_sq, not_retrieved = entsoe_historic_demand(**year_start_end)
+
+    if not_retrieved:
+        logger.warning("Demand data from entsoe could not be retrieved.")
+        # check for generation backup from former runs
+        file_path = Path(
+            "./", "entsoe_data", f"load_entsoe_{scn_name}.csv"
+        ).resolve()
+        if os.path.isfile(file_path):
+            load_sq_backup = pd.read_csv(file_path, index_col="Index")
+            # check for missing columns in backup (former runs)
+            c_backup = [
+                c for c in load_sq_backup.columns if c in not_retrieved
+            ]
+            # remove columns, if found in backup
+            not_retrieved = [c for c in not_retrieved if c not in c_backup]
+            if c_backup:
+                load_sq = pd.concat(
+                    [load_sq, load_sq_backup.loc[:, c_backup]], axis=1
+                )
+                logger.info(f"Appended data from former runs for {c_backup}")
+        save_entsoe_data(load_sq, file_path=file_path)
+
+    if not_retrieved:
+        logger.warning(
+            f"Backup data of 2019 is used instead for {not_retrieved}"
         )
-        load_sq = pd.read_csv(
+        load_sq_backup = pd.read_csv(
             "data_bundle_egon_data/entsoe/load_entsoe.csv", index_col="Index"
+        )
+        load_sq_backup.index = load_sq.index
+        load_sq = pd.concat(
+            [load_sq, load_sq_backup.loc[:, not_retrieved]], axis=1
         )
 
     # Delete existing data
@@ -1700,7 +1820,7 @@ def insert_loads_sq(scn_name="status2019"):
     session = sessionmaker(bind=engine)()
 
     # get the corresponding bus per foreign country
-    entsoe_to_bus = entsoe_to_bus_etrago()
+    entsoe_to_bus = entsoe_to_bus_etrago(scn_name)
 
     # Calculate and insert demand timeseries per etrago bus_id
     for country in load_sq.columns:
@@ -1733,7 +1853,27 @@ if "eGon2035" in config.settings()["egon-data"]["--scenarios"]:
     insert_per_scenario.update([tyndp_generation, tyndp_demand])
 
 if "status2019" in config.settings()["egon-data"]["--scenarios"]:
-    insert_per_scenario.update([insert_generators_sq, insert_loads_sq])
+    insert_per_scenario.update(
+        [
+            wrapped_partial(
+                insert_generators_sq, scn_name="status2019", postfix="_19"
+            ),
+            wrapped_partial(
+                insert_loads_sq, scn_name="status2019", postfix="_19"
+            ),
+        ]
+    )
+if "status2023" in config.settings()["egon-data"]["--scenarios"]:
+    insert_per_scenario.update(
+        [
+            wrapped_partial(
+                insert_generators_sq, scn_name="status2023", postfix="_23"
+            ),
+            wrapped_partial(
+                insert_loads_sq, scn_name="status2023", postfix="_23"
+            ),
+        ]
+    )
 
 tasks = tasks + (insert_per_scenario,)
 
@@ -1742,7 +1882,7 @@ class ElectricalNeighbours(Dataset):
     def __init__(self, dependencies):
         super().__init__(
             name="ElectricalNeighbours",
-            version="0.0.10",
+            version="0.0.11",
             dependencies=dependencies,
             tasks=tasks,
         )